changed the test_utils.py file into a test package with seperate files for each function
